New SearchBar Ui and Event model. New SearchView and SearchViewModel logic to accommodate to the SearchBar new models logic and Two new SearchScreen views.
Description:
Now we have 2 new model for the SearchTopFieldBar: SearchBarUiState & SearchBarEvents this way is easier to handle all the events that the searchbar has and is different states.
Now the SearchBar states are handle by the ViewModel to follow single source of truth arch.

We also added 2 new states for displaying empty list when the user have no locations saved or when there is no location found when searching.

We also remove BasicUiState from being extended from SearchListUiState because we are going to be handle loading, offline and error states through the SearchBar

Issues: