Interaction Between Search Result and Weather Screen
Description:
Now when we can tap on the supported locations that we receive as results from our search query and send us back to the WeatherScreen with that locations data. Currently is just stale data so it can just be the list that we feed from the repo.

New argument for WeatherScreen that is set in the AppNavigation and is sent from our SearchScreen

Issues: