Bug Fix: Clear and refresh search list after saving a location to preventing UI stale/old data
Description:

Explanation: With this changes the bug where the UI would keep the old state from the list item is fix. What was happening is that essentially because the item was update `saveTag` to true, it is view as a new item (which it is in technically) and the Flow and Set would not remove it so we had to clear the `_cachedSupportedLocations` list every time there is a fetch from the Flow.

- When a location's save status is changed from the modal sheet, the search bar is now cleared and the search results are refreshed.
- Created a new `clearSearch` private function in `SearchViewModel` to encapsulate the logic for clearing the search bar text and re-triggering an empty search.
- The `onLocationSaveStatusChanged` function now calls `clearSearch` to ensure the UI reflects the new saved state immediately.
- Removed unused and commented-out code, including a call to `showSavedLocation` and a test `log` function.
- In `getAllLocationSupported` coroutine, the `_cachedSupportedLocations` list is now explicitly cleared before being repopulated to prevent stale data.
- The `.idea/studiobot.xml` file has been added to `.gitignore`.